Move external module into main file
